Refactor IO check
Move the parsing of iostat output to a separate function so that it can
be tested. Add a test using representative output from iostat on a
Debian and CentOS system (comments show the iostat versions used in
each; notably, the Debian version has two additional fields not in the
CentOS version).

It may be difficult to see in the diff, but this commit doesn't
fundamentally change anything done in the IO check, and the test fails.