Fixed supervisord integration
Broken due to my confusion between event_object and aggregation_key.

@clofresh we renamed event_object into aggregation_key for the API
only and I (wrongly) assumed it would apply to dogstream as
well. Since it does not appear to be the case, I've reverted an
earlier change that suppressed event_object in favor of
aggregation_key. It now makes sure that both are present and equal.

@rl-0x0 my bad for breaking the supervisord integration.