Update FindReplace string checking
Basically added some special cases so we can call the replace and it
will *only* make the replacement if the data it gets is good.
If you get something like:
tags:
as the replacement string, it'll skip it because there's no value. It
will also skip if you send an empty string as the replacement string.