monkeypatch to fix memory leak in memcache library
Instantiating a memcache.Client object will create a memcache._Host
object that stores the debuglog method of the Client object. That
_Host object gets stored in the original Client object, presumably
causing that memory to get lost in time and space forever. The
workaround is to set Client.debuglog to None so that _Host won't
try to store anything, avoiding the cyclic reference. (#278)