Fixes #1131 by lowering loglevel when stopping
When stopping the agent, supervisord sends SIGTERM to both the
forwarder and dogstatsd at the same time. There is a race condition
issue, when the forwarder stops before dogstatsd, the latter will try
to flush the metrics to the forwarder in its loop, causing an exception.
This is expected as we're stopping, so if the stop flag is set just
consider this exception as INFO.