Batch raw OID retrieval to bypass 'tooBig' SNMP error
When trying to poll about 40 OIDs at the same time we got
an SNMP 'tooBig' error and no stats were exported at all.
One can change the size limit in the server's SNMP configuration
but it can be done only so much.
This is a bit naive approach to fix it which seems to be working.