Disable tagging by hostname by default for HAProxy service check.
In general you care about a {backend,service} combination but not the
HAProxy host itself. With tagging by host, the default setup for an
HAProxy monitor will trigger N errors everytime a backend goes down,
where N is the number of haproxy load balancers.

This will make things less noisy but keeps an option to enable the
tagging by hostname if you would like it.

I also updated the formatting on the YAML file to match the HTTP
check so it's more readable/clear.