Move to new docker-based CI Travis builds
The builds must not rely on sudo from now on.
That means either fetching pre-compiled binaries from
upstream or fetching the sources and compiling directly
on the build machine.
However, as the released upstream sources do not change
we can benefit from the caching system and cache the
binaries once we have them.
All these software are in $INTEGRATIONS_DIR

Introduction also of running the integration tests against
different versions of the software described in the build
matrix.

Also a bit of documentation on how to use the new test system.

+ Several more fixes of the test suite codebase