Always sent OK status with Postgres check, even with cached connection.
We were incorrectly only sending an OK status when a connection was
first established. This means any monitors on this check would be in
a NODATA state after the first run of the check.

Now we'll send an OK once we have a working connection, even if the
connection had to be reset.