[postgres] Fix custom metrics collection
Fix regression introduced by #1395.
The idea is that because the pre-processing function
`_process_customer_metrics`
was modifying "in-place" instance attributes, and for instance
converting the string "GAUGE" to the method "AgentCheck.gauge" at the
first pass, the second time it would expect to find a string and fail
when it encounters a method.
Instead we cache the pre-preprocessed dict the first time, and hit this
dictionnary in the next runs if it is present.