[ci] Use a bucket to download all tarballs :racehorse:
This shouldn't be used too often because we have another layer of
caching (depending on dd-agent-travis-cache).
Although, when we really have to download the source and recompile it's
often when SourceForge or GitHub have an outage. By placing every egg in
the same basket we ensure that it's *our fault* if things go wrong.

Also:
* upgraded postgres versions so had to update the tests for that