Don't require psutil to check if a PID exists
The `Daemon` class got an additional dependency with `psutil` with this
commit 21ba11a9d23.
The problem is that the core of the agent now depends on psutil, and for
instance the source install requires much more stuff to get an agent
working properly than before (build-essential, python-dev). For that
reason, psutil should be an optional requirement.