[ci/cache] cache depends on the ci/*.rb file version
We add a md5 hash of the flavor file ci/*.rb to the slug enabling a
cache per version of the flavor file which allows you to change a flavor
without having to clear the cache.
A downside is that we'll probably need at one point a job to clean the
bucket, deleting old untouched tarballs.