[redis] Fixed slowlog-max-len issue & added tests
A wrong variable name was used somewhere in our redis check. It had
never been pointed out because this piece of code was executed only when
the "slowlog-max-len" parameter is used, which wasn't done during our
tests.

It has been fixed and an additionnal test has been written to test this
piece of code as well.

The slowlog tests have also been slightly
modified so that they run fster and do not rely on a slowlog threshold
that may not be reached on some powerful hardware.