[systemd] Package updated to be systemd compliant
The install script has been modified (it uses updated-rc.d to restart
the agent, which will implicitly call systemd if available) and the
datadog-agent.service systemd descriptor has been added as part of the
package (so that omnibus puts it in the .deb packages).

This service declaration is basically pluging supervisord into systemd,
nothing more (as shown here
https://gist.github.com/tonyseek/ddc7e3a89f0d78e685d6 for instance)

The supervisord start call has been ut in a separate script.
It allows us to have access to the PATH (/bin/sh is called to run the
script) while keeping a one-level forking scheme thanks to exec (/bin/sh
gets replaced by supervisord that will then fork and launch the
different components of our agent. Being a direct child of systemd, the
forked processes are too).