[systemd] Package updated to be systemd compliant
The install script has been modified (it uses updated-rc.d to restart
the agent, which will implicitly call systemd if available) and the
datadog-agent.service systemd descriptor has been added as part of the
package (so that omnibus puts it in the .deb packages).

This service declaration is basically pluging supervisord into systemd,
nothing more (as shown here
https://gist.github.com/tonyseek/ddc7e3a89f0d78e685d6 for instance)

The weird ExecStart directive is unfortunately necessary. Since systemd
itself doesn't have access to environment variable such as PATH
(because it doesn't load .bashrc scripts or equivalents) we are forced
to run our start command into a shell script to be able to override the
$PATH. That being said, supervisord runs as a daemon which makes the
original /bin/bash command exit and supervisord as well as  all the agent's
processes end up as direct child from the main systemd daemon, which is
the expected behaviour (croissants for you if you managed to read the
previous sentence without blinking :) )