[deps] Add a `pip-allow-failures` wrapper script
Takes a requirements file as an argument.
It will do:
`$PIP_COMMAND install $PIP_OPTIONS dependency`
for all the dependencies specified in the requirements file.
By default `PIP_COMMAND=pip` and no `PIP_OPTIONS` are specified.

It's useful to be able to skip PyPI dependencies if we don't really need
them. I placed it in a separate script, it's not ideal but it avoids
copying over to multiple places (Rakefile, source install, CI).