[install script] fix final health check
[skip ci]
Before that, if the forwarder was not in a ready state at the first
try, the installation script would fail directly without trying a few
more times as we used a trap on error and set -e.

This commit clears the trap and use set +e to allow a few failures for
the status verification.