[process] add a PID cache, avoid iter on all procs
The new init_config `pid_cache_duration` is the period during which the
agent won't try to refresh the "matching" PID list for each instance.
It avoids having to iterate over all the processes every run, the cache
is invalidated when:
* the deadline expire
* or one of the cached process can't be read anymore

One would want to lower the cache deadline if he wants to alert on
no-data.