[http_check] fix type cast of http status code opt
PyYaml was returning an int for the expected status code when
"http_response_status_code: 200" (for instance) was used in the yaml
config while a Regex (and therefore an string) is expected.

Added a line to the http_check.yaml.conf.example file to document the
possibility for the user to use a Regex if he wants to define a set of
"CONNECTION_OK" HTTP status codes (and not only a single one).