[GCE] Add config option to use new gce hostname
The current hostname we use is not unique, if people have several
projects with instances with the same name, their hostname will be
the same, their tags will be wrong and their metrics will be aggre-
gated.
This commit fixes that, and adds an option on the config to keep
the backward compatibility.