[elastic] Add shard-level metrics and serv. check
This adds shard-level metrics to elasticsearch. The feature is disabled
by default. A flag has to be set in elastic.yaml to enable it.

It fetches, for every shard (primary or replica) of every index, a set
of metrics restricted to those shards. It also create a `state` service
check on a per-shard basis. Metrics and state are fetched from the
`/_stats?level=shard` endpoint. A test has also been added.

The metrics and service checks are sent with a bunch of new tags:
`es_node:node_name`, `es_shard:shard_name`, `es_index:index_name`,
`es_role:(primary|replica)` to make it easy to define a scope to
monitor/graph in our backend. The `shard_specific` tag is also added to
avoid confusion with other stats metrics, aggregated on elasticsearch's
end.