Allow for more complex object hiearchy in debug.get_check
In the original version of get_check(name,config_str),
if the class specified by name was not a direct
subclass of AgentCheck, the get_check function would return the
direct subclass of AgentCheck. This verison takes into account the object
hierarchy and returns the lowest class on the object hierarchy instead.
Examples for this are HTTPCheck(NetworkCheck(AgentCheck))