[network] normalize instance_name tags
Instance name used to be a part of the check name and hence normalized
on the agent before being sent to Datadog.
As part of changes #1642 which introduced support for custom tags,
'instance_name' is now sent as a proper tag. Normalize it, to avoid mismatch and
backward incompatiblity.

Fix #1796.