[windows] Refactor Windows service layer
The service part of the windows agent has been rewritten: instead of
having the service file (former `agent.py`) including the "supervisor"
logic, both now live in two separate files.

In addition to a better separation of concerns, this allows us to put
our supervisor outside of the executable generated by `py2exe`. Which
again, makes it easier for somebody willing to contribute to access the
agent's python venv and it's pip. Hacking on the agent on Windows should
now be a bit easier.

The `agent_service.py` file takes care of managing the global `Datadog
Agent` service... all it does is starting  `supervisor.py` using
`subprocess.Popen`) and killing that process when the OS asks the
service to stop.

It comes along with the set of changes brought by the Omnibus 4 build.