[logging] Avoid initializing logging multiple times
The initialization of logging is usually done at the beginning
of a python module. This can lead the init of logging to happen
multiple times when a module that has already initialized it
imports another module that also inits it, which causes the
logs of the first module to be duplicated in the second module's
logs.

For instance agent.py imports jmxfetch.py, which causes the collector's
logs to also end up in jmxfetch.py's logs.

I don't think we ever want the init of logging to happen multiple times
per process, hence this commit.