[windows] Use sockets in our supervisor + path fix
Before the supervisor was killed in a very brutal and rude manner.
Now both layers can talk through a socket bound to port 9001 so that
the service layer can ask the supervisor to terminate smoothly and
the latter replies once he's done so that the service can finally
terminate.

Also, the way the service layer computes the path of the supervisor has
been enhanced to support different agent installation configurations:

So far, only the MSI installer was supported. With this change (along
with the rest of the Omnibus 4 build), it is now possible to use the
source install script as well (even though we don't have one yet on
Windows). Finally, just git pull-ing the `dd-agent` repo, installing
the requirements and running the agent (YES, even as a regular
Windows service) has also been made possible.

This commit is particularly intented to people willing to get hack on
the agent.