[process] Improve sampling of cpu.pct
Improvements:
- Non-blocking call to `cpu_percent` after the first sample
- Retrieves an averaged value since the last call (instead of an
averaged value over the fraction of a second defined by
`cpu_check_interval`)

To use the non-blocking call we keep Process instances through
multiple check runs.

Also updated a test on the check, on which we stop using real PIDs
because the test was previously relying on the of `find_pids` to
return a list of identical PIDs, whereas `find_pids` actually returns
a set of PIDs.