[core] improve detection of agent process from pid
Sometimes the agent fails to restart after a "hard" reboot because the pidfile
contains the pid of a non-datadog running process, and the only check the agent
did before this commit was to check if the pid was associated with a running
process.

This commit implements a new function `is_my_process` which verifies on Linux if the
`psutil.Process.cmdline` ("command line this process has been called with") is
the good one (corresponding to the current executable). This function is used
instead of `pid_exists` when `psutil` is installed and checks the pid contained
in the pidfile, thus solving the above-mentioned problem.