[go_expvar] add configuration option for custom metric namespace
This change adds the ability to configure a custom namespace for your `go_expvar` metrics. By default, the old value (`go_expvar`) is used. The mock tests were updated to confirm that the new behavior works as expected.

In addition to the changes around the check script, the Go utility the tests use was modified as well. A manual GC trigger was added to ensure garbage collection has run at least once.