[jenkins] add support for reporting build parameters as event tags
Jenkins checks can now be configured to report arbitrary build parameters as
event tags. For example, consider a config that looks like

```
init_config:

instances:
  - name: default
    jenkins_home: /var/lib/jenkins
    parameters:
      GIT_BRANCH: branch
```

and a Jenkins job where the `GIT_BRANCH` parameter is set to `master`. With
this patch the resulting DataDog event will have a `branch:master` tag
associated with it.

(This particular example illustrates why Stripe authored this patch: we don't
use the Jenkins Git plugin but still want branch information to be threaded
through to DataDog.)

While expanding `test_jenkins.py` to cover this new functionality, I replaced
the dict-to-XML helper with handwritten XML strings because it seemed simpler
than enhancing the helper to generate the more complex XML needed for the new
test coverage.