[windows][requests] Bundle default CA certs of `requests`
py2exe only includes the python files of the python packages in the
`library` zipped file. This causes the `requests` library to be shipped
without CA certs, so all the https requests made with it fail (unless
SSL cert validation is explicitely disabled).

This fixes the issue by:
1. shipping the default certificates of `requests` with the Windows
agent as an extra data file
2. overriding the path to the default CA bundle that `requests` uses
(which is done in quite an ugly way, but I haven't found any better
way)

[skip ci]