Add a check to monitor mesos task cpu and mem usage
`mesos_task.py` runs against a mesos slave instance, and reports metrics
on tasks running on that instance.

Included metrics are allocated cpus and mem for the task, and the actual
cpu and mem in use for the task.

Task metrics are tagged with the executing framework. Their names are
normalized for known frameworks (mesos and chronos for now).