[source] log properly (package style) & cleanup: part 1
[olivier.vielpeau@datadoghq.com] To ease the release of these changes,
this commit requires another commit that applies the changes to the
source install script.

Previsouly logging was done using supervisor & stdout, totally different
from the package logging. There are options in the configuration to
change the log location, so let's use them instead of doing this trick!

In order to do this, this changes were made:
* remove supervisor processes logging (except `supervisor` itself)
* add a step to configure log location in the source install script

In order to clean weird conf location & others:
* move `supervisor.conf` from `supervisord/` to `agent/` with
`datadog.conf`
* update `bin/agent` to reflect this change
* remove `bin/info` as it is handled in `bin/agent` now (was probably
still here for _legacy_ reasons)
* put launchd logs in the same directory as other logs (`logs/`)