[centos][init] Wait for supervisord to stop before `kill -9`ing it
Without this fix, the `stop` command (-> `killproc`) would only wait for
0.1 second before `kill -9`ing supervisord, so if the collector for
instance took more than that to stop, a subsequent `start` would
fail because we'd try to start a collector before the old one had died.

This fix uses an option of `killproc` so that it waits for 30 seconds
(similar to our init script on debian) before `kill -9`ing
supervisord (which is plenty of time since supervisord `kill -9`s
the processes that haven't stopped after 10 seconds, by default).

[skip ci]