Make max number of consul services configurable
This change adds a mechanism to increase the number of consul services
that can be queries by the agent: Currently, without this change, the
culling mechanism in consul.py would remove any services (even
whitelisted ones) in excess of the first 50. With this change, you can
set `max_services` in consul.yaml, per consul agent.

Further, to optionally remove any limit on the number of services, this
makes it possible to set `max_services` to a negative number disable
service culling entirely.

This change also adds tests for the new negative-maximum behavior and
straightens out the docs in consul.yaml.example.