[http_check] SNI support for cert_expiration
Previously SNI wasn't supported for the cert_expiration part of the check,
because we were not passing the hostname to wrap_socket.

It can only be done with a SSLContext. The code is taken from the Python docs
and should act exactly the same as before.

Doc: https://docs.python.org/2.7/library/ssl.html#ssl.SSLContext.wrap_socket