Update ceph.py
Add to ceph_args two parameters in order to take the name of the Ceph cluster, since If you set the cluster name at install time, then the ceph_args is not going to take the valid name and failed. It needs to change _collect_raw and check functions in order to pass the name parameter. Also it would be great to pass the name of the cluster on the yaml file.