[python_rq] added python-rq AgentCheck to collect queues and workers metrics
Python RQ uses Redis as a backing service to store jobs states each
time one of them is launched. Users may define their own queues
and may choose how many workers are assigned to each queue.

The AgentCheck collects metrics so that developers are aware of
the queues usage, while knowing how much their workers are busy. This
is important to decide when it's the time to spawn more workers or
to assign some workers only to particular queues.