[haproxy] Add `collate_status_tags_per_host` option
Collates status tags to one of (`available`, `unavailable`).

We don't send aggregate statuses when this new option is enabled.
With the option enabled, we already send the collated status
with more fine-grained tags than the aggregated statuses.
As a consequence the metric becomes hard to use on graphs/monitors
if the aggregated statuses are also sent, so let's avoid sending
them in this case.

Also:
- Fix aggregates when `count_status_by_service` is disabled
- Refactor `count_per_status` to use an actual counter