[wmi][test] reload after mocking 🔁
If the `checks.libs.wmi.sampler` module is loaded before the
`tests.core.test_wmi` test suite runs, the mocking defined in the
`setUp` fixture method is never loaded.
Systematically reload the `checks.libs.wmi.sampler` module in
`tests.core.test_wmi`'s `setUp` fixture to avoid this.

[olivier.vielpeau@datadoghq.com: Only apply the mocking and reload the
module in the test class that needs it, because reloading in the base
class messes with the the wmi_check tests]