[haproxy] Fix `KeyError` when an unknown status is found
An unknown status (i.e. one that's not listed in
`Services.ALL_STATUSES`) would make the check fail with a `KeyError`
because that status wasn't in the initialized `dict` of statuses.

With this commit the check now sends unknown statuses as:
- an `UNAVAILABLE` status if `collate_status_tags_per_host` is enabled,
in order to keep the number of contexts low
- the unknown status as-is if `collate_status_tags_per_host` is left
disabled, in order to give maximum visibility in the actual statuses