[elastic] tag stats metric with the node name 🏷 (#2696)
When the `cluster_stats` parameter is enabled in the user configuration,
the Datadog Agent's Elasticsearch check attempts to resolve and
attach with the corresponding metrics, the hostname of the different
nodes.

On Amazon Elasticsearch v1.5.2, the node objects
returned by the `_cluster/nodes/stats?all=true` endpoint does not return
this information. Instead the following keys are available:
```
root@xxxx:~# at node_stats.json| jq '.nodes[] | select(.name ==
"${hostname}") | keys
[
  "breakers",
  "fs",
  "indices",
  "jvm",
  "name",
  "os",
  "process",
  "thread_pool",
  "timestamp"
]
```

`name` is a unique piece of identity. It generally match the hostname.
Use it an additional tag with the namespace `node_name`.