[postgres] Allow check to use UNIX socket as host
The PostgreSQL check has been changed to try to connect to the
PostgreSQL database via a UNIX socket if the hostname provided in the
configuration begins with a forward-slash ('/').

This behaviour is similar to the 'psql' command-line utility. From
psql(1):

  --host=hostname
      Specifies the host name of the machine on which the server is
      running. If the value begins with a slash, it is used as the
      directory for the Unix-domain socket.

With the difference being that we interpret it as an exact path, whereas
psql interprets it as a directory to look for a socket with a
particularly formatted name.