[service_discovery][etcd] Use a correct index to determine last update
The `res.etcd_index` value that was previously used is the
`X-Etcd-Index` header of etcd's http response.

According to etcd's docs this index is the current index of the whole
cluster, so any modification on any key increments this value.

Using it as a reference to determine if the datastore has changed
would trigger a config reload on every modification in the datastore
(whereas we want a config reload only when the `sd_template_dir`
changes).

To fix this, we parse the `modifiedIndex` of all the returned keys
instead, and we take the max index as the correct index.