[wmi_check][test] Fix flaky `tag_queries` test (#2751)
`CommandLine` is generally empty for the `svchost` processes, which
makes the check fail when used as a target property of `tag_queries`
(for good reasons IMO). Let's use a property that we're sure exists
and is not empty.

Also, make the error message clearer when the target property exists
but is empty.