[collector][logging] Log check starts with info level on first run only
Since we removed the LaconicFilter (see #2605) the collector is filled
with `Running check xxxx` messages.

Let's log it at the `info` level only on the first run (since it can
be useful to quickly detect if a check is consistently running too
slowly), but only log it at the `debug` level afterwards.