[debug] improve `run_check` compatibility
Datadog Agent checks do not necessarily inherit directly from the
`AgentCheck` class. When it is the case, the `run_check` DEBUG method
would fail loading the check.
This happens with WMI checks for instance.

Allow the `run_check` DEBUG method to load any agent check that inherits
from the `AgentCheck`.