[ci] do not run integration tests if not relevant
This PR tries to avoid running a integration test for an agent check (and
only this, core integration tests, linter and core mocked tests are
always run) when it is not required.

To do this, we first lists all modified files, and then
- check if a check file was modified (in `checks.d`)
- check if a check test file was modified
- ignore fixtures and configuration

If any file other than the one listed above are modified, all tests will
be run. Otherwise, a list of modfied checks is built, and only these
tests will be run (with the core tests).

This only concerns Travis and is disabled on Appveyor.