[ceph] Fix check failure when health detail contains non-osd info
The `detail` key of `health detail` can contain info on the health of
non-osd units, for example:

```
mon.vagrant-ubuntu-trusty-64 low disk space -- 18% avail
```

In that case the check would previously fail with an IndexError because
the regex we use wouldn't match this string.

Ignore the entry instead.

Includes a test case that reproduces the issue.