[ceph] Improve health metrics (#2852)
* Catch `KeyError` on `osd_pool_stats`

We do it for all the other keys of the raw dict, so let's do it for
this one too (even though I'm not a huge fan of this "try/except
everything" approach).

* Update health metrics for usability

Behavior before this commit: the `ceph.num_near_full_osds` and
`ceph.num_full_osds` would either:
* take the value `0` and be tagged with no `osd` tag if no osd is
reporting health issues
* take a value representing the usage percentage, tagged by `osd`

This doesn't really make sense with respect to the name of the metrics
(`num_*`). To solve this, replace these metrics with the following:
* `ceph.num_near_full_ods` and `ceph.num_full_osds` report the total
number of osds that are respectively near full and full. Not tagged
by osd.
* when some osds report health issues, the check sends a
`ceph.osd.pct_used` metric which reports the usage
percentage, tagged by osd. Unfortunately we can't send `0` values
on this metric when no osd reports health issues since we can't tag by
osd in that case.

This should make these metrics more usable. Also, use `gauge` since
there's no reason to use `count`.