Make args optional in kafka_consumer.py
The consumer_groups, topics, and partitions arguments are each now optional. Any omitted values will be fetched from Zookeeper. If a value is omitted, the parent value must still be it's expected type (typically a dict).

This was originally a bug fix of https://github.com/DataDog/dd-agent/pull/2781/ but quickly turned grew into a much bigger refactor in order to handle various bugs and clarify the logic used to fetch missing information from Zookeeper. This does not add the ability to fetch consumer offsets from Kafka.

Modern Kafka consumers store their offset in Kafka rather than Zookeeper, but support for this has not been added yet. It's tricky because this check is much simpler if it assumes a single place can be queried for all consumer offsets, but currently there's no easy way to do that. Once KIP-88 is implemented, it will be much easier to add this functionality, although it would only work for Kafka brokers >= 0.10.2.0. In the meantime, you can instrument your individual kafka consumers to submit their offsets to Datadog.