[install_script][yum] Enable all the repos when installing `datadog-agent`
The yum package of `datadog-agent` has a dependency on `initscripts`.
If we disable all the repos when installing `datadog-agent` _and_ if
`initscripts` is not installed `yum` won't be able to find the dep,
and the install will fail.

To fix this we fall back to leaving all the repos enabled when
installing the agent.

[skip ci]