[collector][emitter] Ignore all non-unicode-decodable characters
We've seen cases (on Japanese and Korean versions of Windows) where
the json encoding would fail on a UnicodeDecodeError because of
non-utf8-decodable characters in strings.

To work around this, when this happens, we make another pass on the
whole payload to remove all the non-decodable characters and log the
affected strings.