[Mongodb] Check was sending password in the clear if the password includes (#2940)
a character that needs to be URL encoded (i.e. '/').  Account for
passwords that include URL encoded chars

Also allow "_" to be a username in the regular expression matching
username/password in a URI.  Otherwise, a username with a "_" will not
match, and the password will be sent in the clear