[core][subprocess output] Do not set close_fds=True
This was introduced by
https://github.com/DataDog/dd-agent/commit/24cf7823c254925ab061baf6aff71
2678a4c793f

But it’s not needed anymore and as a result the collector is looping
over thousand of file descriptors, trying to close them (most of the
time it cannot) and use a lot of CPU to do so. This commit change that
and revert to the default `close_fds=False`

Also I’m removing a bunch of options that are not used anywhere.