[core] display expected service check count (#2938)
We add one service check per check, to see whether the check is broken
or not. The info command then shows a misleading number service checks.

Let's fix it and not count the additional service check.
`max` is there as an additional security, the current code should never
need it but we never know.