[ci] Fix unjustified Travis test skips (#2995)
And add some debug logging.

Using `TRAVIS_COMMIT` on the `git diff` command sometimes fail because
git can't find the object reference for some reason. Instead use the
latest commit of the PR branch and use a difference comparison to only
take into account the changes brought by the PR branch since the latest
common commit with the base branch.

Also, pin the version of a gem dependency since later versions are
incompatible with the provided ruby on Travis (1.9.3).