[psutil] Only set `psutil.PROCFS_PATH` once in the collector
Since it's a global setting of `psutil` we should only set it once for
the whole collector.

To preserve backwards compatibility in the process check configuration
we only deprecate the setting in its `init_config` (instead of removing
it entirely), and use this setting when it doesn't conflict with what's
specified in `datadog.conf`.