[ci] Fix Travis jobs timing out (#3017)
* [postgres][ci] Fix user that connects to CI postgres db

* [ci] Do run the cleanup step on Travis

For some reason some jobs on Travis wouldn't exit
unless the processes that are spawned are all killed

* [ci] Improve reliability on Travis

- Run FluentD as a daemon
- Increase rabbitmq timeout

* [ci][apache] Fix cleanup

If we remove the conf during the `before_cache` stage we can't stop
apache afterwards. Instead move the conf file out of the cached dir
and move it back in apache's dir before stopping.