Postfix check should pass raise_on_empty_output=False
A recent fix to subprocess_output.py in dd-agent 5.10.0 exposed an
issue in the postfix check. This fix added robustness to the way that
subproces_output.py determines whether or not the output is empty, and
since the postfix check produces empty output when evaluating empty
queues, this caused an error to be raised.

By passing raise_on_empty_output=False we can assert that empty output
is acceptable for the postfix check, and receive metrics (rather than
raising an error).