Export `http_host` as a tag for PHP-FPM metrics
The current PHP-FPM mechanism provides no mechanism to associate a
metric with the `http_host` that was specified to enact the check.

This can be useful in circumstances when a single server has multiple
FPM pools.  Notably this data can be discerned from the extant
`pool` tag, however it is quite possible that the pool name and the host
header can be entirely different.